Refactor: split large files into smaller components
CopilotService (2274 → 6 partial class files):
- CopilotService.cs (~925 lines) — core lifecycle, session mgmt
- CopilotService.Organization.cs (~240 lines) — groups, pinning, sorting
- CopilotService.Persistence.cs (~292 lines) — disk I/O, save/load
- CopilotService.Bridge.cs (~189 lines) — remote mode, bridge sync
- CopilotService.Events.cs (~345 lines) — SDK event handling
- CopilotService.Utilities.cs (~320 lines) — git, badges, models, history

Dashboard.razor (1342 → 1221 lines):
- Extract ExpandedSessionView.razor (~231 lines) — full chat view
  with 16 Parameters + 12 EventCallbacks

SessionSidebar.razor (756 → ~680 lines):
- Extract SessionListItem.razor (~110 lines) — single session row
  with context menu, rename, pin/move/close actions

Zero behavior changes. All 99 tests pass.
