+++ AutoPilot.App/Services/CopilotService.cs
@@ -6,7 +6,7 @@ using GitHub.Copilot.SDK;
 
 namespace AutoPilot.App.Services;
 
-public class CopilotService : IAsyncDisposable
+public partial class CopilotService : IAsyncDisposable
 {
     private readonly ConcurrentDictionary<string, SessionState> _sessions = new();
     private readonly ChatDatabase _chatDb;
@@ -246,93 +246,6 @@ public class CopilotService : IAsyncDisposable
     }
 
     /// <summary>
-    /// Initialize in Remote mode: connect WsBridgeClient for state-sync with server.
-    /// </summary>
-    private async Task InitializeRemoteAsync(ConnectionSettings settings, CancellationToken ct)
-    {
-        var wsUrl = settings.RemoteUrl!.TrimEnd('/');
-        if (wsUrl.StartsWith("https://", StringComparison.OrdinalIgnoreCase))
-            wsUrl = "wss://" + wsUrl[8..];
-        else if (wsUrl.StartsWith("http://", StringComparison.OrdinalIgnoreCase))
-            wsUrl = "ws://" + wsUrl[7..];
-        else
-            wsUrl = "wss://" + wsUrl;
-
-        Debug($"Remote mode: connecting to {wsUrl}");
-
-        // Wire WsBridgeClient events to our events
-        _bridgeClient.OnStateChanged += () =>
-        {
-            SyncRemoteSessions();
-            InvokeOnUI(() => OnStateChanged?.Invoke());
-        };
-        _bridgeClient.OnContentReceived += (s, c) =>
-        {
-            // Update local session history from remote events
-            var session = GetRemoteSession(s);
-            if (session != null)
-            {
-                var existing = session.History.LastOrDefault(m => m.IsAssistant && !m.IsComplete);
-                if (existing != null)
-                    existing.Content += c;
-                else
-                    session.History.Add(new ChatMessage("assistant", c, DateTime.Now, ChatMessageType.Assistant) { IsComplete = false });
-            }
-            InvokeOnUI(() => OnContentReceived?.Invoke(s, c));
-        };
-        _bridgeClient.OnToolStarted += (s, tool, id) =>
-        {
-            var session = GetRemoteSession(s);
-            session?.History.Add(ChatMessage.ToolCallMessage(tool, id));
-            InvokeOnUI(() => OnToolStarted?.Invoke(s, tool, id, null));
-        };
-        _bridgeClient.OnToolCompleted += (s, id, result, success) =>
-        {
-            var session = GetRemoteSession(s);
-            var toolMsg = session?.History.LastOrDefault(m => m.ToolCallId == id);
-            if (toolMsg != null)
-            {
-                toolMsg.IsComplete = true;
-                toolMsg.IsSuccess = success;
-                toolMsg.Content = result;
-            }
-            InvokeOnUI(() => OnToolCompleted?.Invoke(s, id, result, success));
-        };
-        _bridgeClient.OnReasoningReceived += (s, rid, c) => InvokeOnUI(() => OnReasoningReceived?.Invoke(s, rid, c));
-        _bridgeClient.OnReasoningComplete += (s, rid) => InvokeOnUI(() => OnReasoningComplete?.Invoke(s, rid));
-        _bridgeClient.OnIntentChanged += (s, i) => InvokeOnUI(() => OnIntentChanged?.Invoke(s, i));
-        _bridgeClient.OnUsageInfoChanged += (s, u) => InvokeOnUI(() => OnUsageInfoChanged?.Invoke(s, u));
-        _bridgeClient.OnTurnStart += (s) =>
-        {
-            var session = GetRemoteSession(s);
-            if (session != null) session.IsProcessing = true;
-            InvokeOnUI(() => OnTurnStart?.Invoke(s));
-        };
-        _bridgeClient.OnTurnEnd += (s) =>
-        {
-            var session = GetRemoteSession(s);
-            if (session != null)
-            {
-                session.IsProcessing = false;
-                // Mark last assistant message as complete
-                var lastAssistant = session.History.LastOrDefault(m => m.IsAssistant && !m.IsComplete);
-                if (lastAssistant != null) lastAssistant.IsComplete = true;
-            }
-            InvokeOnUI(() => OnTurnEnd?.Invoke(s));
-        };
-        _bridgeClient.OnSessionComplete += (s, sum) => InvokeOnUI(() => OnSessionComplete?.Invoke(s, sum));
-        _bridgeClient.OnError += (s, e) => InvokeOnUI(() => OnError?.Invoke(s, e));
-
-        await _bridgeClient.ConnectAsync(wsUrl, settings.RemoteToken, ct);
-
-        IsInitialized = true;
-        IsRemoteMode = true;
-        NeedsConfiguration = false;
-        Debug("Connected to remote server via WebSocket bridge");
-        OnStateChanged?.Invoke();
-    }
-
-    /// <summary>
     /// Initialize in Demo mode: wire up DemoService events for local mock responses.
     /// </summary>
     private void InitializeDemo()
@@ -383,102 +296,6 @@ public class CopilotService : IAsyncDisposable
         Debug("Demo mode initialized");
         OnStateChanged?.Invoke();
     }
-
-    /// <summary>
-    /// Sync remote session list from WsBridgeClient into our local _sessions dictionary.
-    /// </summary>
-    private void SyncRemoteSessions()
-    {
-        var remoteSessions = _bridgeClient.Sessions;
-        var remoteActive = _bridgeClient.ActiveSessionName;
-
-        Debug($"SyncRemoteSessions: {remoteSessions.Count} remote sessions, active={remoteActive}");
-
-        // Add/update sessions from remote
-        foreach (var rs in remoteSessions)
-        {
-            if (!_sessions.ContainsKey(rs.Name))
-            {
-                Debug($"SyncRemoteSessions: Adding session '{rs.Name}'");
-                var info = new AgentSessionInfo
-                {
-                    Name = rs.Name,
-                    Model = rs.Model,
-                    CreatedAt = rs.CreatedAt,
-                    SessionId = rs.SessionId,
-                    WorkingDirectory = rs.WorkingDirectory,
-                    GitBranch = GetGitBranch(rs.WorkingDirectory),
-                };
-                _sessions[rs.Name] = new SessionState
-                {
-                    Session = null!,  // No local CopilotSession in remote mode
-                    Info = info
-                };
-            }
-            // Update processing state from server
-            if (_sessions.TryGetValue(rs.Name, out var state))
-            {
-                state.Info.IsProcessing = rs.IsProcessing;
-                state.Info.MessageCount = rs.MessageCount;
-            }
-        }
-
-        // Remove sessions that no longer exist on server
-        var remoteNames = remoteSessions.Select(s => s.Name).ToHashSet();
-        foreach (var name in _sessions.Keys.ToList())
-        {
-            if (!remoteNames.Contains(name))
-                _sessions.TryRemove(name, out _);
-        }
-
-        // Sync history from WsBridgeClient cache
-        // Don't overwrite if local history has messages not yet reflected by server
-        var sessionsNeedingHistory = new List<string>();
-        foreach (var (name, messages) in _bridgeClient.SessionHistories)
-        {
-            if (_sessions.TryGetValue(name, out var s))
-            {
-                if (messages.Count >= s.Info.History.Count)
-                {
-                    Debug($"SyncRemoteSessions: Syncing {messages.Count} messages for '{name}'");
-                    s.Info.History.Clear();
-                    s.Info.History.AddRange(messages);
-                }
-            }
-        }
-
-        // Request history for sessions that have messages but no local history yet
-        foreach (var rs in remoteSessions)
-        {
-            if (rs.MessageCount > 0 && _sessions.TryGetValue(rs.Name, out var s) && s.Info.History.Count == 0
-                && !_bridgeClient.SessionHistories.ContainsKey(rs.Name))
-            {
-                sessionsNeedingHistory.Add(rs.Name);
-            }
-        }
-
-        if (sessionsNeedingHistory.Count > 0)
-        {
-            _ = Task.Run(async () =>
-            {
-                foreach (var name in sessionsNeedingHistory)
-                {
-                    try { await _bridgeClient.RequestHistoryAsync(name); }
-                    catch { }
-                }
-            });
-        }
-
-        // Sync active session — only on first load, not on every update
-        // (user may have selected a different session locally on mobile)
-        if (_activeSessionName == null && remoteActive != null && _sessions.ContainsKey(remoteActive))
-            _activeSessionName = remoteActive;
-
-        Debug($"SyncRemoteSessions: Done. _sessions has {_sessions.Count} entries, active={_activeSessionName}");
-    }
-
-    private AgentSessionInfo? GetRemoteSession(string name) =>
-        _sessions.TryGetValue(name, out var state) ? state.Info : null;
 
     /// <summary>
     /// Disconnect from current client and reconnect with new settings
@@ -546,331 +363,6 @@ public class CopilotService : IAsyncDisposable
             }),
             _ => new CopilotClient()
         };
-    }
-
-    /// <summary>
-    /// Gets a list of persisted session GUIDs from ~/.copilot/session-state
-    /// </summary>
-    public IEnumerable<PersistedSessionInfo> GetPersistedSessions()
-    {
-        // In remote mode, return persisted sessions from the bridge
-        if (IsRemoteMode)
-        {
-            return _bridgeClient.PersistedSessions
-                .Select(p => new PersistedSessionInfo
-                {
-                    SessionId = p.SessionId,
-                    Title = p.Title,
-                    Preview = p.Preview,
-                    WorkingDirectory = p.WorkingDirectory,
-                    LastModified = p.LastModified,
-                });
-        }
-
-        if (!Directory.Exists(SessionStatePath))
-            return Enumerable.Empty<PersistedSessionInfo>();
-
-        return Directory.GetDirectories(SessionStatePath)
-            .Select(dir => new DirectoryInfo(dir))
-            .Where(di => Guid.TryParse(di.Name, out _))
-            .Select(di => CreatePersistedSessionInfo(di))
-            .OrderByDescending(s => s.LastModified);
-    }
-
-    private PersistedSessionInfo CreatePersistedSessionInfo(DirectoryInfo di)
-    {
-        string? title = null;
-        string? preview = null;
-        string? workingDir = null;
-
-        var eventsFile = Path.Combine(di.FullName, "events.jsonl");
-        if (File.Exists(eventsFile))
-        {
-            try
-            {
-                // Read first few lines to find first user message and working directory
-                foreach (var line in File.ReadLines(eventsFile).Take(50))
-                {
-                    if (string.IsNullOrWhiteSpace(line)) continue;
-                    using var doc = JsonDocument.Parse(line);
-                    var root = doc.RootElement;
-                    
-                    if (!root.TryGetProperty("type", out var typeEl)) continue;
-                    var type = typeEl.GetString();
-
-                    // Get working directory from session.start
-                    if (type == "session.start" && workingDir == null)
-                    {
-                        if (root.TryGetProperty("data", out var data))
-                        {
-                            // Try data.context.cwd first (newer format), then data.workingDirectory
-                            if (data.TryGetProperty("context", out var ctx) &&
-                                ctx.TryGetProperty("cwd", out var cwd))
-                            {
-                                workingDir = cwd.GetString();
-                            }
-                            else if (data.TryGetProperty("workingDirectory", out var wd))
-                            {
-                                workingDir = wd.GetString();
-                            }
-                        }
-                    }
-                    
-                    // Get first user message
-                    if (type == "user.message" && title == null)
-                    {
-                        if (root.TryGetProperty("data", out var data) &&
-                            data.TryGetProperty("content", out var content))
-                        {
-                            preview = content.GetString();
-                            if (!string.IsNullOrEmpty(preview))
-                            {
-                                // Create truncated title (max 60 chars)
-                                title = preview.Length > 60 
-                                    ? preview[..57] + "..." 
-                                    : preview;
-                                // Clean up newlines for title
-                                title = title.Replace("\n", " ").Replace("\r", "");
-                            }
-                        }
-                        break; // Got what we need
-                    }
-                }
-            }
-            catch { /* Ignore parse errors */ }
-        }
-
-        // Use events.jsonl modification time for accurate "last used" sorting
-        var eventsFileInfo = new FileInfo(eventsFile);
-        var lastUsed = eventsFileInfo.Exists ? eventsFileInfo.LastWriteTime : di.LastWriteTime;
-
-        // Priority: alias > active session name > first message > "Untitled session"
-        var alias = GetSessionAlias(di.Name);
-        string resolvedTitle;
-        if (!string.IsNullOrEmpty(alias))
-            resolvedTitle = alias;
-        else if (title != null)
-            resolvedTitle = title;
-        else
-        {
-            var activeMatch = _sessions.Values.FirstOrDefault(s => s.Info.SessionId == di.Name);
-            resolvedTitle = activeMatch?.Info.Name ?? "Untitled session";
-        }
-
-        return new PersistedSessionInfo
-        {
-            SessionId = di.Name,
-            LastModified = lastUsed,
-            Path = di.FullName,
-            Title = resolvedTitle,
-            Preview = preview ?? "No preview available",
-            WorkingDirectory = workingDir
-        };
-    }
-
-    /// <summary>
-    /// Check if a session was still processing when the app last closed
-    /// </summary>
-    private bool IsSessionStillProcessing(string sessionId)
-    {
-        var eventsFile = Path.Combine(SessionStatePath, sessionId, "events.jsonl");
-        if (!File.Exists(eventsFile)) return false;
-
-        try
-        {
-            string? lastLine = null;
-            foreach (var line in File.ReadLines(eventsFile))
-            {
-                if (!string.IsNullOrWhiteSpace(line))
-                    lastLine = line;
-            }
-            if (lastLine == null) return false;
-
-            using var doc = JsonDocument.Parse(lastLine);
-            var type = doc.RootElement.GetProperty("type").GetString();
-            
-            var activeEvents = new[] { 
-                "assistant.turn_start", "tool.execution_start", 
-                "tool.execution_progress", "assistant.message_delta",
-                "assistant.reasoning", "assistant.reasoning_delta",
-                "assistant.intent"
-            };
-            return activeEvents.Contains(type);
-        }
-        catch { return false; }
-    }
-
-    /// <summary>
-    /// Get the last tool name and assistant message from events.jsonl for status display
-    /// </summary>
-    private (string? lastTool, string? lastContent) GetLastSessionActivity(string sessionId)
-    {
-        var eventsFile = Path.Combine(SessionStatePath, sessionId, "events.jsonl");
-        if (!File.Exists(eventsFile)) return (null, null);
-
-        try
-        {
-            string? lastTool = null;
-            string? lastContent = null;
-
-            foreach (var line in File.ReadLines(eventsFile))
-            {
-                if (string.IsNullOrWhiteSpace(line)) continue;
-                using var doc = JsonDocument.Parse(line);
-                var root = doc.RootElement;
-                var type = root.GetProperty("type").GetString();
-
-                if (type == "tool.execution_start" && root.TryGetProperty("data", out var toolData))
-                {
-                    if (toolData.TryGetProperty("toolName", out var tn))
-                        lastTool = tn.GetString();
-                }
-                else if (type == "assistant.message" && root.TryGetProperty("data", out var msgData))
-                {
-                    if (msgData.TryGetProperty("content", out var content))
-                    {
-                        var c = content.GetString();
-                        if (!string.IsNullOrEmpty(c))
-                            lastContent = c;
-                    }
-                }
-            }
-            return (lastTool, lastContent);
-        }
-        catch { return (null, null); }
-    }
-
-    /// <summary>
-    /// Load conversation history from events.jsonl
-    /// </summary>
-    private List<ChatMessage> LoadHistoryFromDisk(string sessionId)
-    {
-        var history = new List<ChatMessage>();
-        var eventsFile = Path.Combine(SessionStatePath, sessionId, "events.jsonl");
-        
-        if (!File.Exists(eventsFile))
-            return history;
-
-        try
-        {
-            // Track tool calls by ID so we can update them when complete
-            var toolCallMessages = new Dictionary<string, ChatMessage>();
-
-            foreach (var line in File.ReadLines(eventsFile))
-            {
-                if (string.IsNullOrWhiteSpace(line)) continue;
-                
-                using var doc = JsonDocument.Parse(line);
-                var root = doc.RootElement;
-                
-                if (!root.TryGetProperty("type", out var typeEl)) continue;
-                var type = typeEl.GetString();
-                
-                if (!root.TryGetProperty("data", out var data)) continue;
-                var timestamp = DateTime.Now;
-                if (root.TryGetProperty("timestamp", out var tsEl))
-                    DateTime.TryParse(tsEl.GetString(), out timestamp);
-
-                switch (type)
-                {
-                    case "user.message":
-                    {
-                        if (data.TryGetProperty("content", out var userContent))
-                        {
-                            var content = userContent.GetString();
-                            if (!string.IsNullOrEmpty(content))
-                            {
-                                var msg = ChatMessage.UserMessage(content);
-                                msg.Timestamp = timestamp;
-                                history.Add(msg);
-                            }
-                        }
-                        break;
-                    }
-
-                    case "assistant.message":
-                    {
-                        // Add reasoning if present
-                        if (data.TryGetProperty("reasoningText", out var reasoningEl))
-                        {
-                            var reasoning = reasoningEl.GetString();
-                            if (!string.IsNullOrEmpty(reasoning))
-                            {
-                                var msg = ChatMessage.ReasoningMessage("restored");
-                                msg.Content = reasoning;
-                                msg.IsComplete = true;
-                                msg.IsCollapsed = true;
-                                msg.Timestamp = timestamp;
-                                history.Add(msg);
-                            }
-                        }
-
-                        // Add assistant text content (skip if only tool requests with no text)
-                        if (data.TryGetProperty("content", out var assistantContent))
-                        {
-                            var content = assistantContent.GetString()?.Trim();
-                            if (!string.IsNullOrEmpty(content))
-                            {
-                                var msg = ChatMessage.AssistantMessage(content);
-                                msg.Timestamp = timestamp;
-                                history.Add(msg);
-                            }
-                        }
-                        break;
-                    }
-
-                    case "tool.execution_start":
-                    {
-                        var toolName = data.TryGetProperty("toolName", out var tn) ? tn.GetString() ?? "" : "";
-                        var toolCallId = data.TryGetProperty("toolCallId", out var tc) ? tc.GetString() : null;
-                        
-                        // Skip report_intent — it's noise in history
-                        if (toolName == "report_intent") break;
-
-                        // Extract tool input if available
-                        string? inputStr = null;
-                        if (data.TryGetProperty("input", out var inputEl))
-                            inputStr = inputEl.ToString();
-                        else if (data.TryGetProperty("arguments", out var argsEl))
-                            inputStr = argsEl.ToString();
-
-                        var msg = ChatMessage.ToolCallMessage(toolName, toolCallId, inputStr);
-                        msg.Timestamp = timestamp;
-                        history.Add(msg);
-                        if (toolCallId != null)
-                            toolCallMessages[toolCallId] = msg;
-                        break;
-                    }
-
-                    case "tool.execution_complete":
-                    {
-                        var toolCallId = data.TryGetProperty("toolCallId", out var tc) ? tc.GetString() : null;
-                        if (toolCallId != null && toolCallMessages.TryGetValue(toolCallId, out var msg))
-                        {
-                            msg.IsComplete = true;
-                            msg.IsSuccess = data.TryGetProperty("success", out var s) && s.GetBoolean();
-                            msg.IsCollapsed = true;
-
-                            if (data.TryGetProperty("result", out var result))
-                            {
-                                // Prefer detailedContent, fall back to content
-                                var content = result.TryGetProperty("detailedContent", out var dc) ? dc.GetString() : null;
-                                if (string.IsNullOrEmpty(content) && result.TryGetProperty("content", out var c))
-                                    content = c.GetString();
-                                msg.Content = content ?? "";
-                            }
-                        }
-                        break;
-                    }
-                }
-            }
-        }
-        catch
-        {
-            // Ignore parse errors, return what we have
-        }
-
-        return history;
     }
 
     /// <summary>
@@ -1094,342 +586,6 @@ ALWAYS run the relaunch script as the final step after making changes to this pr
         ReconcileOrganization();
         OnStateChanged?.Invoke();
         return info;
-    }
-
-    private static readonly HashSet<string> FilteredTools = new() { "report_intent", "skill", "store_memory" };
-
-    private void HandleSessionEvent(SessionState state, SessionEvent evt)
-    {
-        state.HasReceivedEventsSinceResume = true;
-        var sessionName = state.Info.Name;
-        void Invoke(Action action)
-        {
-            if (_syncContext != null)
-                _syncContext.Post(_ => action(), null);
-            else
-                action();
-        }
-        
-        switch (evt)
-        {
-            case AssistantReasoningEvent reasoning:
-                Invoke(() =>
-                {
-                    OnReasoningReceived?.Invoke(sessionName, reasoning.Data.ReasoningId ?? "", reasoning.Data.Content ?? "");
-                });
-                break;
-
-            case AssistantReasoningDeltaEvent reasoningDelta:
-                Invoke(() =>
-                {
-                    OnReasoningReceived?.Invoke(sessionName, reasoningDelta.Data.ReasoningId ?? "", reasoningDelta.Data.DeltaContent ?? "");
-                });
-                break;
-
-            case AssistantMessageDeltaEvent delta:
-                var deltaContent = delta.Data.DeltaContent;
-                state.HasReceivedDeltasThisTurn = true;
-                state.CurrentResponse.Append(deltaContent);
-                Invoke(() => OnContentReceived?.Invoke(sessionName, deltaContent ?? ""));
-                break;
-
-            case AssistantMessageEvent msg:
-                var msgContent = msg.Data.Content;
-                var msgId = msg.Data.MessageId;
-                // Deduplicate: SDK fires this event multiple times for resumed sessions
-                if (!string.IsNullOrEmpty(msgContent) && !state.HasReceivedDeltasThisTurn && msgId != state.LastMessageId)
-                {
-                    state.LastMessageId = msgId;
-                    state.CurrentResponse.Append(msgContent);
-                    state.Info.LastUpdatedAt = DateTime.Now;
-                    Invoke(() => OnContentReceived?.Invoke(sessionName, msgContent));
-                }
-                break;
-
-            case ToolExecutionStartEvent toolStart:
-                if (toolStart.Data == null) break;
-                var startToolName = toolStart.Data.ToolName ?? "unknown";
-                var startCallId = toolStart.Data.ToolCallId ?? "";
-                var toolInput = ExtractToolInput(toolStart.Data);
-                if (!FilteredTools.Contains(startToolName))
-                {
-                    // Flush any accumulated assistant text before adding tool message
-                    FlushCurrentResponse(state);
-                    
-                    var toolMsg = ChatMessage.ToolCallMessage(startToolName, startCallId, toolInput);
-                    state.Info.History.Add(toolMsg);
-                    
-                    Invoke(() =>
-                    {
-                        OnToolStarted?.Invoke(sessionName, startToolName, startCallId, toolInput);
-                        OnActivity?.Invoke(sessionName, $"🔧 Running {startToolName}...");
-                    });
-                }
-                else if (state.CurrentResponse.Length > 0)
-                {
-                    // Separate text blocks around filtered tools so they don't run together
-                    state.CurrentResponse.Append("\n\n");
-                }
-                break;
-
-            case ToolExecutionCompleteEvent toolDone:
-                if (toolDone.Data == null) break;
-                var completeCallId = toolDone.Data.ToolCallId ?? "";
-                var completeToolName = toolDone.Data?.GetType().GetProperty("ToolName")?.GetValue(toolDone.Data)?.ToString();
-                var resultStr = FormatToolResult(toolDone.Data.Result);
-                var hasError = toolDone.Data.Error != null;
-
-                // Skip filtered tools
-                if (completeToolName != null && FilteredTools.Contains(completeToolName))
-                    break;
-                if (resultStr == "Intent logged")
-                    break;
-
-                // Update the matching tool message in history
-                var histToolMsg = state.Info.History.LastOrDefault(m => m.ToolCallId == completeCallId);
-                if (histToolMsg != null)
-                {
-                    histToolMsg.IsComplete = true;
-                    histToolMsg.IsSuccess = !hasError;
-                    histToolMsg.Content = resultStr;
-                }
-
-                Invoke(() =>
-                {
-                    OnToolCompleted?.Invoke(sessionName, completeCallId, resultStr, !hasError);
-                    OnActivity?.Invoke(sessionName, hasError ? "❌ Tool failed" : "✅ Tool completed");
-                });
-                break;
-
-            case ToolExecutionProgressEvent:
-                Invoke(() => OnActivity?.Invoke(sessionName, "⚙️ Tool executing..."));
-                break;
-
-            case AssistantIntentEvent intent:
-                var intentText = intent.Data.Intent ?? "";
-                Invoke(() =>
-                {
-                    OnIntentChanged?.Invoke(sessionName, intentText);
-                    OnActivity?.Invoke(sessionName, $"💭 {intentText}");
-                });
-                break;
-
-            case AssistantTurnStartEvent:
-                state.HasReceivedDeltasThisTurn = false;
-                Invoke(() =>
-                {
-                    OnTurnStart?.Invoke(sessionName);
-                    OnActivity?.Invoke(sessionName, "🤔 Thinking...");
-                });
-                break;
-
-            case AssistantTurnEndEvent:
-                Invoke(() =>
-                {
-                    OnTurnEnd?.Invoke(sessionName);
-                    OnActivity?.Invoke(sessionName, "");
-                });
-                break;
-
-            case SessionIdleEvent:
-                CompleteResponse(state);
-                // Refresh git branch — agent may have switched branches
-                state.Info.GitBranch = GetGitBranch(state.Info.WorkingDirectory);
-                break;
-
-            case SessionStartEvent start:
-                state.Info.SessionId = start.Data.SessionId;
-                Debug($"Session ID assigned: {start.Data.SessionId}");
-                SaveActiveSessionsToDisk();
-                break;
-
-            case SessionUsageInfoEvent usageInfo:
-                var uData = usageInfo.Data;
-                var uModel = uData?.GetType().GetProperty("Model")?.GetValue(uData)?.ToString();
-                var uCurrentTokens = uData?.GetType().GetProperty("CurrentTokens")?.GetValue(uData) as int?;
-                var uTokenLimit = uData?.GetType().GetProperty("TokenLimit")?.GetValue(uData) as int?;
-                var uInputTokens = uData?.GetType().GetProperty("InputTokens")?.GetValue(uData) as int?;
-                var uOutputTokens = uData?.GetType().GetProperty("OutputTokens")?.GetValue(uData) as int?;
-                if (!string.IsNullOrEmpty(uModel))
-                    state.Info.Model = uModel;
-                Invoke(() => OnUsageInfoChanged?.Invoke(sessionName, new SessionUsageInfo(uModel, uCurrentTokens, uTokenLimit, uInputTokens, uOutputTokens)));
-                break;
-
-            case AssistantUsageEvent assistantUsage:
-                var aData = assistantUsage.Data;
-                var aModel = aData?.GetType().GetProperty("Model")?.GetValue(aData)?.ToString();
-                var aInput = aData?.GetType().GetProperty("InputTokens")?.GetValue(aData) as int?;
-                var aOutput = aData?.GetType().GetProperty("OutputTokens")?.GetValue(aData) as int?;
-                if (!string.IsNullOrEmpty(aModel))
-                    state.Info.Model = aModel;
-                if (aInput.HasValue || aOutput.HasValue)
-                {
-                    Invoke(() => OnUsageInfoChanged?.Invoke(sessionName, new SessionUsageInfo(aModel, null, null, aInput, aOutput)));
-                }
-                break;
-
-            case SessionErrorEvent err:
-                var errMsg = err.Data?.Message ?? "Unknown error";
-                Invoke(() => OnError?.Invoke(sessionName, errMsg));
-                state.ResponseCompletion?.TrySetException(new Exception(errMsg));
-                state.Info.IsProcessing = false;
-                Invoke(() => OnStateChanged?.Invoke());
-                break;
-
-            case SessionModelChangeEvent modelChange:
-                var newModel = modelChange.Data?.NewModel;
-                if (!string.IsNullOrEmpty(newModel))
-                {
-                    state.Info.Model = newModel;
-                    Debug($"Session '{sessionName}' model changed to: {newModel}");
-                    Invoke(() => OnUsageInfoChanged?.Invoke(sessionName, new SessionUsageInfo(newModel, null, null, null, null)));
-                    Invoke(() => OnStateChanged?.Invoke());
-                }
-                break;
-                
-            default:
-                break;
-        }
-    }
-
-    private static string FormatToolResult(object? result)
-    {
-        if (result == null) return "";
-        if (result is string str) return str;
-        try
-        {
-            var resultType = result.GetType();
-            // Prefer DetailedContent (has richer info like file paths) over Content
-            foreach (var propName in new[] { "DetailedContent", "detailedContent", "Content", "content", "Message", "message", "Text", "text", "Value", "value" })
-            {
-                var prop = resultType.GetProperty(propName);
-                if (prop != null)
-                {
-                    var val = prop.GetValue(result)?.ToString();
-                    if (!string.IsNullOrEmpty(val)) return val;
-                }
-            }
-            var json = JsonSerializer.Serialize(result, new JsonSerializerOptions { WriteIndented = true });
-            if (json != "{}" && json != "null") return json;
-        }
-        catch { }
-        return result.ToString() ?? "";
-    }
-
-    private static string? ExtractToolInput(object? data)
-    {
-        if (data == null) return null;
-        try
-        {
-            var type = data.GetType();
-            // Try common property names for tool input/arguments
-            foreach (var propName in new[] { "Input", "Arguments", "Args", "Parameters", "input", "arguments" })
-            {
-                var prop = type.GetProperty(propName);
-                if (prop == null) continue;
-                var val = prop.GetValue(data);
-                if (val == null) continue;
-                if (val is string s && !string.IsNullOrEmpty(s)) return s;
-                try
-                {
-                    var json = JsonSerializer.Serialize(val, new JsonSerializerOptions { WriteIndented = false });
-                    if (json != "{}" && json != "null" && json != "\"\"") return json;
-                }
-                catch { return val.ToString(); }
-            }
-        }
-        catch { }
-        return null;
-    }
-
-    private void TryAttachImages(MessageOptions options, List<string> imagePaths)
-    {
-        try
-        {
-            var attachments = new List<UserMessageDataAttachmentsItem>();
-            foreach (var path in imagePaths)
-            {
-                if (!File.Exists(path)) continue;
-                var fileItem = new UserMessageDataAttachmentsItemFile
-                {
-                    Path = path,
-                    DisplayName = System.IO.Path.GetFileName(path)
-                };
-                attachments.Add(fileItem);
-            }
-
-            if (attachments.Count == 0) return;
-            options.Attachments = attachments;
-            Debug($"Attached {attachments.Count} image(s) via SDK");
-        }
-        catch (Exception ex)
-        {
-            Debug($"Failed to attach images via SDK: {ex.Message}");
-        }
-    }
-
-    /// <summary>Flush accumulated assistant text to history without ending the turn.</summary>
-    private void FlushCurrentResponse(SessionState state)
-    {
-        var text = state.CurrentResponse.ToString();
-        if (string.IsNullOrEmpty(text)) return;
-        
-        var msg = new ChatMessage("assistant", text, DateTime.Now);
-        state.Info.History.Add(msg);
-        state.Info.MessageCount = state.Info.History.Count;
-        
-        if (!string.IsNullOrEmpty(state.Info.SessionId))
-            _ = _chatDb.AddMessageAsync(state.Info.SessionId, msg);
-        
-        state.CurrentResponse.Clear();
-        state.HasReceivedDeltasThisTurn = false;
-    }
-
-    private void CompleteResponse(SessionState state)
-    {
-        if (!state.Info.IsProcessing) return; // Already completed (e.g. timeout)
-        
-        var response = state.CurrentResponse.ToString();
-        if (!string.IsNullOrEmpty(response))
-        {
-            var msg = new ChatMessage("assistant", response, DateTime.Now);
-            state.Info.History.Add(msg);
-            state.Info.MessageCount = state.Info.History.Count;
-
-            // Write-through to DB
-            if (!string.IsNullOrEmpty(state.Info.SessionId))
-                _ = _chatDb.AddMessageAsync(state.Info.SessionId, msg);
-        }
-        state.ResponseCompletion?.TrySetResult(response);
-        state.CurrentResponse.Clear();
-        state.Info.IsProcessing = false;
-        state.Info.LastUpdatedAt = DateTime.Now;
-        OnStateChanged?.Invoke();
-        
-        // Fire completion notification
-        var summary = response.Length > 100 ? response[..100] + "..." : response;
-        OnSessionComplete?.Invoke(state.Info.Name, summary);
-        IncrementBadge();
-
-        // Auto-dispatch next queued message
-        if (state.Info.MessageQueue.Count > 0)
-        {
-            var nextPrompt = state.Info.MessageQueue[0];
-            state.Info.MessageQueue.RemoveAt(0);
-            _ = Task.Run(async () =>
-            {
-                try
-                {
-                    await Task.Delay(500);
-                    await SendPromptAsync(state.Info.Name, nextPrompt);
-                }
-                catch (Exception ex)
-                {
-                    Debug($"Failed to send queued message: {ex.Message}");
-                    OnError?.Invoke(state.Info.Name, $"Queued message failed: {ex.Message}");
-                }
-            });
-        }
     }
 
     public async Task<string> SendPromptAsync(string sessionName, string prompt, List<string>? imagePaths = null, CancellationToken cancellationToken = default)
@@ -1718,332 +874,6 @@ ALWAYS run the relaunch script as the final step after making changes to this pr
 
     public int SessionCount => _sessions.Count;
 
-    #region Session Organization (groups, pinning, sorting)
-
-    public void LoadOrganization()
-    {
-        try
-        {
-            if (File.Exists(OrganizationFile))
-            {
-                var json = File.ReadAllText(OrganizationFile);
-                Organization = JsonSerializer.Deserialize<OrganizationState>(json) ?? new OrganizationState();
-            }
-            else
-            {
-                Organization = new OrganizationState();
-            }
-        }
-        catch (Exception ex)
-        {
-            Debug($"Failed to load organization: {ex.Message}");
-            Organization = new OrganizationState();
-        }
-
-        // Ensure default group always exists
-        if (!Organization.Groups.Any(g => g.Id == SessionGroup.DefaultId))
-        {
-            Organization.Groups.Insert(0, new SessionGroup
-            {
-                Id = SessionGroup.DefaultId,
-                Name = SessionGroup.DefaultName,
-                SortOrder = 0
-            });
-        }
-
-        ReconcileOrganization();
-    }
-
-    public void SaveOrganization()
-    {
-        try
-        {
-            var json = JsonSerializer.Serialize(Organization, new JsonSerializerOptions { WriteIndented = true });
-            File.WriteAllText(OrganizationFile, json);
-        }
-        catch (Exception ex)
-        {
-            Debug($"Failed to save organization: {ex.Message}");
-        }
-    }
-
-    /// <summary>
-    /// Ensure every active session has a SessionMeta entry and clean up orphans.
-    /// </summary>
-    private void ReconcileOrganization()
-    {
-        var activeNames = _sessions.Keys.ToHashSet();
-        bool changed = false;
-
-        // Add missing sessions to default group
-        foreach (var name in activeNames)
-        {
-            if (!Organization.Sessions.Any(m => m.SessionName == name))
-            {
-                Organization.Sessions.Add(new SessionMeta
-                {
-                    SessionName = name,
-                    GroupId = SessionGroup.DefaultId
-                });
-                changed = true;
-            }
-        }
-
-        // Fix sessions pointing to deleted groups
-        var groupIds = Organization.Groups.Select(g => g.Id).ToHashSet();
-        foreach (var meta in Organization.Sessions)
-        {
-            if (!groupIds.Contains(meta.GroupId))
-            {
-                meta.GroupId = SessionGroup.DefaultId;
-                changed = true;
-            }
-        }
-
-        // Remove metadata for sessions that no longer exist
-        int removed = Organization.Sessions.RemoveAll(m => !activeNames.Contains(m.SessionName));
-        if (removed > 0) changed = true;
-
-        if (changed) SaveOrganization();
-    }
-
-    public void PinSession(string sessionName, bool pinned)
-    {
-        var meta = Organization.Sessions.FirstOrDefault(m => m.SessionName == sessionName);
-        if (meta != null)
-        {
-            meta.IsPinned = pinned;
-            SaveOrganization();
-            OnStateChanged?.Invoke();
-        }
-    }
-
-    public void MoveSession(string sessionName, string groupId)
-    {
-        var meta = Organization.Sessions.FirstOrDefault(m => m.SessionName == sessionName);
-        if (meta != null && Organization.Groups.Any(g => g.Id == groupId))
-        {
-            meta.GroupId = groupId;
-            SaveOrganization();
-            OnStateChanged?.Invoke();
-        }
-    }
-
-    public SessionGroup CreateGroup(string name)
-    {
-        var group = new SessionGroup
-        {
-            Id = Guid.NewGuid().ToString(),
-            Name = name,
-            SortOrder = Organization.Groups.Max(g => g.SortOrder) + 1
-        };
-        Organization.Groups.Add(group);
-        SaveOrganization();
-        OnStateChanged?.Invoke();
-        return group;
-    }
-
-    public void RenameGroup(string groupId, string name)
-    {
-        var group = Organization.Groups.FirstOrDefault(g => g.Id == groupId);
-        if (group != null)
-        {
-            group.Name = name;
-            SaveOrganization();
-            OnStateChanged?.Invoke();
-        }
-    }
-
-    public void DeleteGroup(string groupId)
-    {
-        if (groupId == SessionGroup.DefaultId) return;
-
-        // Move all sessions in this group to default
-        foreach (var meta in Organization.Sessions.Where(m => m.GroupId == groupId))
-        {
-            meta.GroupId = SessionGroup.DefaultId;
-        }
-
-        Organization.Groups.RemoveAll(g => g.Id == groupId);
-        SaveOrganization();
-        OnStateChanged?.Invoke();
-    }
-
-    public void ToggleGroupCollapsed(string groupId)
-    {
-        var group = Organization.Groups.FirstOrDefault(g => g.Id == groupId);
-        if (group != null)
-        {
-            group.IsCollapsed = !group.IsCollapsed;
-            SaveOrganization();
-            OnStateChanged?.Invoke();
-        }
-    }
-
-    public void SetSortMode(SessionSortMode mode)
-    {
-        Organization.SortMode = mode;
-        SaveOrganization();
-        OnStateChanged?.Invoke();
-    }
-
-    public void SetSessionManualOrder(string sessionName, int order)
-    {
-        var meta = Organization.Sessions.FirstOrDefault(m => m.SessionName == sessionName);
-        if (meta != null)
-        {
-            meta.ManualOrder = order;
-            SaveOrganization();
-        }
-    }
-
-    public void SetGroupOrder(string groupId, int order)
-    {
-        var group = Organization.Groups.FirstOrDefault(g => g.Id == groupId);
-        if (group != null)
-        {
-            group.SortOrder = order;
-            SaveOrganization();
-            OnStateChanged?.Invoke();
-        }
-    }
-
-    /// <summary>
-    /// Returns sessions organized by group, with pinned sessions first and sorted by the current sort mode.
-    /// </summary>
-    public IEnumerable<(SessionGroup Group, List<AgentSessionInfo> Sessions)> GetOrganizedSessions()
-    {
-        var metas = Organization.Sessions.ToDictionary(m => m.SessionName);
-        var allSessions = GetAllSessions().ToList();
-
-        foreach (var group in Organization.Groups.OrderBy(g => g.SortOrder))
-        {
-            var groupSessions = allSessions
-                .Where(s => metas.TryGetValue(s.Name, out var m) && m.GroupId == group.Id)
-                .ToList();
-
-            // Pinned first, then apply sort mode within each partition
-            var sorted = groupSessions
-                .OrderByDescending(s => metas.TryGetValue(s.Name, out var m) && m.IsPinned)
-                .ThenBy(s => ApplySort(s, metas))
-                .ToList();
-
-            yield return (group, sorted);
-        }
-    }
-
-    private object ApplySort(AgentSessionInfo session, Dictionary<string, SessionMeta> metas)
-    {
-        return Organization.SortMode switch
-        {
-            SessionSortMode.LastActive => DateTime.MaxValue - session.LastUpdatedAt,
-            SessionSortMode.CreatedAt => DateTime.MaxValue - session.CreatedAt,
-            SessionSortMode.Alphabetical => session.Name,
-            SessionSortMode.Manual => (object)(metas.TryGetValue(session.Name, out var m) ? m.ManualOrder : int.MaxValue),
-            _ => DateTime.MaxValue - session.LastUpdatedAt
-        };
-    }
-
-    public bool HasMultipleGroups => Organization.Groups.Count > 1;
-
-    public SessionMeta? GetSessionMeta(string sessionName) =>
-        Organization.Sessions.FirstOrDefault(m => m.SessionName == sessionName);
-
-    #endregion
-
-    /// <summary>
-    /// Save active session list to disk so we can restore on relaunch
-    /// </summary>
-    private void SaveActiveSessionsToDisk()
-    {
-        try
-        {
-            var entries = _sessions.Values
-                .Where(s => s.Info.SessionId != null)
-                .Select(s => new ActiveSessionEntry
-                {
-                    SessionId = s.Info.SessionId!,
-                    DisplayName = s.Info.Name,
-                    Model = s.Info.Model
-                })
-                .ToList();
-            
-            var json = JsonSerializer.Serialize(entries, new JsonSerializerOptions { WriteIndented = true });
-            File.WriteAllText(ActiveSessionsFile, json);
-        }
-        catch (Exception ex)
-        {
-            Debug($"Failed to save active sessions: {ex.Message}");
-        }
-    }
-
-    /// <summary>
-    /// Load and resume all previously active sessions
-    /// </summary>
-    public async Task RestorePreviousSessionsAsync(CancellationToken cancellationToken = default)
-    {
-        if (File.Exists(ActiveSessionsFile))
-        {
-            try
-            {
-                var json = await File.ReadAllTextAsync(ActiveSessionsFile, cancellationToken);
-                var entries = JsonSerializer.Deserialize<List<ActiveSessionEntry>>(json);
-                if (entries != null && entries.Count > 0)
-                {
-                    Debug($"Restoring {entries.Count} previous sessions...");
-
-                    foreach (var entry in entries)
-                    {
-                        try
-                        {
-                            // Skip if already active
-                            if (_sessions.ContainsKey(entry.DisplayName)) continue;
-                            
-                            // Check the session still exists on disk
-                            var sessionDir = Path.Combine(SessionStatePath, entry.SessionId);
-                            if (!Directory.Exists(sessionDir)) continue;
-
-                            await ResumeSessionAsync(entry.SessionId, entry.DisplayName, cancellationToken);
-                            Debug($"Restored session: {entry.DisplayName}");
-                        }
-                        catch (Exception ex)
-                        {
-                            Debug($"Failed to restore '{entry.DisplayName}': {ex.Message}");
-
-                            // If the connection broke, recreate the client
-                            if (ex is System.IO.IOException or System.Net.Sockets.SocketException
-                                or ObjectDisposedException
-                                || ex.InnerException is System.IO.IOException or System.Net.Sockets.SocketException
-                                || ex.Message.Contains("Connection", StringComparison.OrdinalIgnoreCase)
-                                || ex.Message.Contains("transport", StringComparison.OrdinalIgnoreCase))
-                            {
-                                Debug("Connection lost during restore, recreating client...");
-                                try
-                                {
-                                    if (_client != null) await _client.DisposeAsync();
-                                    var settings = ConnectionSettings.Load();
-                                    _client = CreateClient(settings);
-                                    await _client.StartAsync(cancellationToken);
-                                    Debug("Client recreated successfully");
-                                }
-                                catch (Exception clientEx)
-                                {
-                                    Debug($"Failed to recreate client: {clientEx.Message}");
-                                    break; // Stop trying to restore sessions
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug($"Failed to load active sessions file: {ex.Message}");
-            }
-        }
-
-    }
-
     public async ValueTask DisposeAsync()
     {
         SaveActiveSessionsToDisk();
@@ -2057,185 +887,6 @@ ALWAYS run the relaunch script as the final step after making changes to this pr
         if (_client != null)
         {
             await _client.DisposeAsync();
-        }
-    }
-
-    public void SaveUiState(string currentPage, string? activeSession = null, int? fontSize = null, string? selectedModel = null)
-    {
-        try
-        {
-            var existing = LoadUiState();
-            var state = new UiState
-            {
-                CurrentPage = currentPage,
-                ActiveSession = activeSession ?? _activeSessionName,
-                FontSize = fontSize ?? existing?.FontSize ?? 20,
-                SelectedModel = selectedModel ?? existing?.SelectedModel
-            };
-            var json = JsonSerializer.Serialize(state);
-            File.WriteAllText(UiStateFile, json);
-        }
-        catch { }
-    }
-
-    public UiState? LoadUiState()
-    {
-        try
-        {
-            if (!File.Exists(UiStateFile)) return null;
-            var json = File.ReadAllText(UiStateFile);
-            return JsonSerializer.Deserialize<UiState>(json);
-        }
-        catch { return null; }
-    }
-    // --- Session Aliases ---
-
-    private Dictionary<string, string>? _aliasCache;
-
-    private Dictionary<string, string> LoadAliases()
-    {
-        if (_aliasCache != null) return _aliasCache;
-        try
-        {
-            if (File.Exists(SessionAliasesFile))
-            {
-                var json = File.ReadAllText(SessionAliasesFile);
-                _aliasCache = JsonSerializer.Deserialize<Dictionary<string, string>>(json) ?? new();
-                return _aliasCache;
-            }
-        }
-        catch { }
-        _aliasCache = new();
-        return _aliasCache;
-    }
-
-    public string? GetSessionAlias(string sessionId)
-    {
-        var aliases = LoadAliases();
-        return aliases.TryGetValue(sessionId, out var alias) ? alias : null;
-    }
-
-    public void SetSessionAlias(string sessionId, string alias)
-    {
-        var aliases = LoadAliases();
-        if (string.IsNullOrWhiteSpace(alias))
-            aliases.Remove(sessionId);
-        else
-            aliases[sessionId] = alias.Trim();
-        _aliasCache = aliases;
-        try
-        {
-            var json = JsonSerializer.Serialize(aliases, new JsonSerializerOptions { WriteIndented = true });
-            File.WriteAllText(SessionAliasesFile, json);
-        }
-        catch { }
-    }
-
-    private static string? GetGitBranch(string? directory)
-    {
-        if (string.IsNullOrEmpty(directory)) return null;
-        try
-        {
-            var headFile = FindGitHead(directory);
-            if (headFile == null) return null;
-            var head = File.ReadAllText(headFile).Trim();
-            return head.StartsWith("ref: refs/heads/")
-                ? head["ref: refs/heads/".Length..]
-                : head.Length >= 8 ? head[..8] : head; // detached HEAD — show short SHA
-        }
-        catch { return null; }
-    }
-
-    private static string? FindGitHead(string dir)
-    {
-        var d = new DirectoryInfo(dir);
-        while (d != null)
-        {
-            var head = Path.Combine(d.FullName, ".git", "HEAD");
-            if (File.Exists(head)) return head;
-            d = d.Parent;
-        }
-        return null;
-    }
-
-    private string? GetSessionWorkingDirectory(string sessionId)
-    {
-        try
-        {
-            var eventsFile = Path.Combine(SessionStatePath, sessionId, "events.jsonl");
-            if (!File.Exists(eventsFile)) return null;
-            // Read only enough lines to find session.start
-            foreach (var line in File.ReadLines(eventsFile).Take(5))
-            {
-                if (string.IsNullOrWhiteSpace(line)) continue;
-                using var doc = JsonDocument.Parse(line);
-                var root = doc.RootElement;
-                if (!root.TryGetProperty("type", out var t) || t.GetString() != "session.start") continue;
-                if (root.TryGetProperty("data", out var data))
-                {
-                    if (data.TryGetProperty("context", out var ctx) &&
-                        ctx.TryGetProperty("cwd", out var cwd))
-                        return cwd.GetString();
-                    if (data.TryGetProperty("workingDirectory", out var wd))
-                        return wd.GetString();
-                }
-            }
-        }
-        catch { }
-        return null;
-    }
-
-    // Dock badge for completed sessions
-    private int _badgeCount;
-
-    private void IncrementBadge()
-    {
-#if MACCATALYST || IOS
-        _badgeCount++;
-        MainThread.BeginInvokeOnMainThread(() =>
-        {
-            try { UIKit.UIApplication.SharedApplication.ApplicationIconBadgeNumber = _badgeCount; }
-            catch { }
-        });
-#endif
-    }
-
-    public void ClearBadge()
-    {
-#if MACCATALYST || IOS
-        _badgeCount = 0;
-        MainThread.BeginInvokeOnMainThread(() =>
-        {
-            try { UIKit.UIApplication.SharedApplication.ApplicationIconBadgeNumber = 0; }
-            catch { }
-        });
-#endif
-    }
-
-    private async Task FetchAvailableModelsAsync()
-    {
-        try
-        {
-            if (_client == null) return;
-            var modelList = await _client.ListModelsAsync();
-            if (modelList != null && modelList.Count > 0)
-            {
-                var models = modelList
-                    .Where(m => !string.IsNullOrEmpty(m.Name))
-                    .Select(m => m.Name!)
-                    .OrderBy(m => m)
-                    .ToList();
-                if (models.Count > 0)
-                {
-                    AvailableModels = models;
-                    Debug($"Loaded {models.Count} models from SDK");
-                    OnStateChanged?.Invoke();
-                }
-            }
-        }
-        catch (Exception ex)
-        {
-            Debug($"Failed to fetch models: {ex.Message}");
         }
     }
 }
